+++ src/main/java/nl/dtls/fairdatapoint/database/mongo/migration/production/Migration_0006_ShapesSharing.java
@@ -1,17 +1,17 @@
 /**
  * The MIT License
  * Copyright © 2017 DTL
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -20,18 +20,32 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package nl.dtls.fairdatapoint.database.mongo.repository;
+package nl.dtls.fairdatapoint.database.mongo.migration.production;
 
-import nl.dtls.fairdatapoint.entity.shape.Shape;
-import org.springframework.data.mongodb.repository.MongoRepository;
+import com.github.mongobee.changeset.ChangeLog;
+import com.github.mongobee.changeset.ChangeSet;
+import com.mongodb.client.MongoCollection;
+import com.mongodb.client.MongoDatabase;
+import com.mongodb.client.model.Filters;
+import com.mongodb.client.model.Updates;
+import nl.dtls.fairdatapoint.Profiles;
+import org.bson.Document;
+import org.springframework.context.annotation.Profile;
 
-import java.util.List;
-import java.util.Optional;
+@ChangeLog(order = "0006")
+@Profile(Profiles.PRODUCTION)
+public class Migration_0006_ShapesSharing {
 
-public interface ShapeRepository extends MongoRepository<Shape, String> {
+    @ChangeSet(order = "0006", id = "Migration_0005_ShapesSharing", author = "migrationBot")
+    public void run(MongoDatabase db) {
+        addShapeDefinitions(db);
+    }
 
-    Optional<Shape> findByUuid(String uuid);
-
-    List<Shape> findAllByPublishedIsTrue();
-
+    private void addShapeDefinitions(MongoDatabase db) {
+        MongoCollection<Document> shapeCol = db.getCollection("shape");
+        shapeCol.updateMany(
+                Filters.exists("published", false),
+                Updates.set("published", false)
+        );
+    }
 }