+++ src/main/java/nl/dtls/fairdatapoint/service/shape/ShapeService.java
@@ -24,7 +24,9 @@ package nl.dtls.fairdatapoint.service.shape;
 
 import nl.dtls.fairdatapoint.api.dto.shape.ShapeChangeDTO;
 import nl.dtls.fairdatapoint.api.dto.shape.ShapeDTO;
+import nl.dtls.fairdatapoint.api.dto.shape.ShapeRemoteDTO;
 import nl.dtls.fairdatapoint.database.mongo.repository.ShapeRepository;
+import nl.dtls.fairdatapoint.entity.exception.ShapeImportException;
 import nl.dtls.fairdatapoint.entity.exception.ValidationException;
 import nl.dtls.fairdatapoint.entity.shape.Shape;
 import nl.dtls.fairdatapoint.entity.shape.ShapeType;
@@ -39,6 +41,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import static java.util.Optional.empty;
 import static java.util.Optional.of;
@@ -58,6 +61,15 @@ public class ShapeService {
 
     public List<ShapeDTO> getShapes() {
         List<Shape> shapes = shapeRepository.findAll();
+        return
+                shapes
+                        .stream()
+                        .map(shapeMapper::toDTO)
+                        .collect(toList());
+    }
+
+    public List<ShapeDTO> getPublishedShapes() {
+        List<Shape> shapes = shapeRepository.findAllByPublishedIsTrue();
         return
                 shapes
                         .stream()
@@ -125,4 +137,20 @@ public class ShapeService {
         return shacl;
     }
 
+    public List<ShapeRemoteDTO> getRemoteShapes(String fdpUrl) {
+        List<ShapeDTO> shapes = ShapeRetrievalUtils.retrievePublishedShapes(fdpUrl);
+        return shapes
+                .stream()
+                .map(s -> shapeMapper.toRemoteDTO(fdpUrl, s))
+                .collect(Collectors.toList());
+    }
+
+    public List<ShapeDTO> importShapes(List<ShapeRemoteDTO> reqDtos) {
+        return
+                reqDtos
+                        .stream()
+                        .map(s -> shapeMapper.fromRemoteDTO(s))
+                        .map(this::createShape)
+                        .collect(Collectors.toList());
+    }
 }